etl/requirements.txt
pandas
httpx
sqlalchemy
psycopg
python-dotenv
dagster
dagster-docker
dagster-postgres
